1. Write functions for joint model of VaR and ES using AL distribution. 2. Add spx500 to the data
