+++ NAMESPACE
@@ -1,17 +1,24 @@
 # Generated by roxygen2: do not edit by hand
 
 export(GetIniParams)
-export(GetIniParamsC)
+export(GetIniParamsAL)
+export(GetIniParamsUni)
+export(GetIniParams_AL)
 export(MidasBetaC)
 export(MidasQuantile)
 export(MixedFreqQuant)
 export(condQuantile)
+export(condVaRES)
 export(objFun)
+export(objFunAL)
 import(dplyr)
 import(pracma)
 importFrom(Rcpp,evalCpp)
 importFrom(Rcpp,sourceCpp)
 importFrom(dfoptim,nmkb)
+importFrom(forecast,Arima)
+importFrom(forecast,auto.arima)
+importFrom(lmtest,coeftest)
 importFrom(minqa,bobyqa)
 importFrom(stats,nlminb)
 importFrom(stats,optim)
